Version 3.5
* Now using `find_library` to find the name for libx11/libx11fixes/libc instead of having it hard coded.
* `open_terminal.py` now closes whenever bash/the program running has been closed.
* Better cleanup in `ipc.py`
* Fixed a bug where `IPC.find_where` will report pid as existsing without checking if it's passes in as a number
* No longer using signals in `os_tools`, instead using `NamedLock`
* Better error handling from `"error"` events generated in `slave.py`
* Fixed `Popen` causing error in `slave.py`
* `slave.py` now cleans up after itself
* Created `ConhostTerminal` which barely works
* `TerminalFrame` now has `_handle_msg_loop`
* Now, `Callable`s can be passed into `TerminalFrame.queue`
* `BaseTerminal` no longer creates zombies/orphans
* `ExpandedExplorer` no longer creates zombies
* `RunManager` no longer relies on `printf` command (only worked on posix)

+++ src/bettertk/terminaltk/terminal.py
@@ -1,19 +1,25 @@
 from __future__ import annotations
 from threading import Thread, Event as _Event
 from time import sleep, perf_counter
-from signal import SIGTERM, SIGKILL
 from subprocess import Popen, PIPE
+from signal import SIGTERM
 import sys
 import os
 
 try:
-    from .ipc import IPC, Event, pid_exists, SELF_PID, SIGUSR1, SIGUSR2
+    from signal import SIGKILL
 except ImportError:
-    from ipc import IPC, Event, pid_exists, SELF_PID, SIGUSR1, SIGUSR2
+    SIGKILL = None
+
+try:
+    from .ipc import IPC, Event, pid_exists, SELF_PID, SIGUSR2, close_all_ipcs
+except ImportError:
+    from ipc import IPC, Event, pid_exists, SELF_PID, SIGUSR2, close_all_ipcs
 from bettertk import IS_WINDOWS, IS_UNIX
 
 
 SLAVE_PATH:str = os.path.join(os.path.dirname(__file__), "slave.py")
+SLAVE_CMD:str = f'{sys.executable} "{SLAVE_PATH}" "{{}}" {{}}'
 
 
 if IS_WINDOWS:
@@ -55,20 +61,29 @@ elif IS_UNIX:
                      "-fs 12 -cu -sb -rightbar -sl 100000 "
     for XTERM_XRM in XTERM_XRMS:
         XTERM_ARGS += f"-xrm 'xterm*{XTERM_XRM}' "
-    XTERM_ARGS += f'-e {sys.executable} "{SLAVE_PATH}" "{{}}" "{{}}"'
-    KONSOLE_ARGS = f'-e {sys.executable} "{SLAVE_PATH}" "{{}}" "{{}}"'
+    XTERM_ARGS:str = XTERM_ARGS.removesuffix(" ")
 else:
     ...
 
 
+def timeout(timeout:float, interval:float, check_exit:Callable[bool]) -> bool:
+    start:float = perf_counter()
+    while perf_counter()-start < timeout:
+        if check_exit():
+            return False
+        sleep(interval)
+    return True
+
+def invert(func:Callable) -> Callable:
+    def inverted(*args:tuple, **kwargs:dict) -> bool:
+        return not func(*args, **kwargs)
+    return inverted
+
 def kill_proc(send_signal:Callable[int,None], is_alive:Callable[bool]) -> None:
     send_signal(SIGTERM)
-    start:float = perf_counter()
-    while perf_counter()-start < 3: # 3 sec for cleanup before SIGKILL
-        if not is_alive():
-            return None
-        sleep(0.05)
-    send_signal(SIGKILL)
+    if timeout(3, 0.05, invert(is_alive)): # 3 sec for cleanup before SIGKILL
+        if SIGKILL is not None:
+            send_signal(SIGKILL)
 
 
 TERMINAL_IPC:IPC = IPC("terminaltk", sig=SIGUSR2)
@@ -76,31 +91,26 @@ TERMINAL_IPC:IPC = IPC("terminaltk", sig=SIGUSR2)
 
 class BaseTerminal:
     __slots__ = "proc", "_running", "resizable", "ipc", "slave_pid", \
-                "_ready_event", "_bindings"
+                "_ready_event", "_bindings", "sep_window"
 
     def __init__(self, *, into:int=None) -> None:
+        self.sep_window:bool = None
         self._bindings:dict = {}
         self.ipc:IPC = TERMINAL_IPC
         self.ipc.find_where("others")
         self.slave_pid:str = None
-        # Create IPC
-        # name:str = IPC.get_empty_name("instance-%d%d%d", "terminaltk",
-        #                               is_abandoned=self._is_abandoned)
-        # if name is None:
-        #     raise FileExistsError("too many terminals exist already")
-        # self.ipc:IPC = IPC(name)
-        # with self.ipc._fs.open("owner_pid", "w") as file:
-        #     file.write(str(SELF_PID))
         # Set up vars
         self.resizable:bool = hasattr(self, "resize")
         self._running:bool = False
         # Start and wait until ready
         self._ready_event:_Event = _Event()
         self.bind("ready", self._ready)
-        self.start(self.ipc.name, into=into)
+        self.start(into=into)
+        assert self.sep_window is not None, "Set `sep_window` please"
         assert self._running, "You must call \"self.run(command, env=env)\" " \
                              'inside "self.start"'
         self._ready_event.wait()
+        Thread(target=self.close_on_dead_slave, daemon=True).start()
 
     def _ready(self, event:Event) -> None:
         self.slave_pid:str = event._from
@@ -109,6 +119,7 @@ class BaseTerminal:
 
     def send_event(self, event:str, *, data:object=None) -> None:
         self.ipc.event_generate(event, where=self.slave_pid, data=data)
+    send = send_event
 
     def bind(self, event:str, handler:Callable[Event,None], **kwargs) -> None:
         # Call handler iff it's from the correct pid or event is ready
@@ -140,20 +151,24 @@ class BaseTerminal:
         return not pid_exists(pid)
 
     def run(self, command:str, *, env:dict[str,str]=os.environ) -> None:
+        slave_cmd:str = SLAVE_CMD.format(self.ipc.name, SELF_PID)
+        command:str = self.str_rreplace_once(command, "%command%", slave_cmd)
         self.proc:Popen = Popen(command, env=env, shell=True, stdout=PIPE,
                                 stderr=PIPE)
         self._running:bool = True
-        while self.slave_pid is None:
-            sleep(0.1)
+        if timeout(2, 0.05, lambda: self.slave_pid is not None):
+            raise RuntimeError("Slave didn't report its pid")
+        def reaper() -> None:
+            self.proc.wait()
+        Thread(target=reaper, daemon=True).start()
 
     def close(self) -> None:
-        _dead_event:_Event = _Event()
-        self.bind("exit", self.ipc.rm_event(_dead_event.set))
-        self.send_event("exit")
-        while (not _dead_event.wait(0.1)) and pid_exists(self.slave_pid):
-            pass
-        if self.proc.poll is None:
-            kill_proc(self.proc.send_signal, lambda: not self.proc.poll())
+        if self.running():
+            self.send("exit")
+        self.proc.wait()
+
+    def close_on_dead_slave(self) -> None:
+        self.proc.wait()
         while self._bindings:
             event, handler = next(iter(self._bindings.keys()))
             self.unbind(event, handler)
@@ -162,33 +177,43 @@ class BaseTerminal:
         return self.proc.poll() is None
 
     def clear(self) -> None:
-        self.send_event("print", data="\r\x1b[2J\x1b[3J\x1b[H\x1bc")
+        self.send("print", data="\r\x1b[2J\x1b[3J\x1b[H\x1bc")
 
     def send_signal(self, signal:int) -> None:
-        self.send_event("signal", data=signal)
+        self.send("signal", data=signal)
 
-    def start(self, ipc_name:str, *, into:int=None):
+    @staticmethod
+    def str_rreplace_once(string:str, find:str, replace:str) -> str:
+        return string.replace(find, replace)
+
+    def start(self, *, into:int=None):
         raise NotImplementedError("Override this method, making sure you " \
                                   "call `self.run(command, env=env)` at " \
-                                  "the end")
+                                  "the end. Also please set `self.sep_window`" \
+                                  "to a boolean.")
 
 
 class XTermTerminal(BaseTerminal):
     __slots__ = ()
 
-    def start(self, ipc_name:str, *, into:tk.Misc=None):
-        args:str = XTERM_ARGS.format(ipc_name, SELF_PID)
-        if into is None:
-            command:str = f"xterm {args}"
-        else:
-            command:str = f"xterm -into {into.winfo_id()} {args}"
+    def start(self, *, into:tk.Misc=None) -> None:
+        self.sep_window:bool = False
+        into_arg:str = ""
+        if into is not None:
+            into_arg:str = f"-into {into.winfo_id()}"
+        command:str = f"xterm {into_arg} {XTERM_ARGS} -e %command%"
         self.run(command, env=os.environ|dict(force_color_prompt="yes"))
-        if XTERM_DEBUG: Thread(target=self.debug_proc_end, daemon=True).start()
 
     def resize(self, *, width:int, height:int) -> None:
         assert isinstance(width, int), "TypeError"
         assert isinstance(height, int), "TypeError"
-        self.send_event("print", data=f"\x1b[4;{height};{width}t")
+        self.send("print", data=f"\x1b[4;{height};{width}t")
+
+
+class DebugXTermTerminal(XTermTerminal):
+    def start(self, *, into:tk.Misc=None) -> None:
+        super().start(into=into)
+        if XTERM_DEBUG: Thread(target=self.debug_proc_end, daemon=True).start()
 
     def debug_proc_end(self) -> None:
         def inner(_:Event=None) -> None:
@@ -202,26 +227,48 @@ class XTermTerminal(BaseTerminal):
         self.bind("exit", inner)
 
 
+class GnomeTermTerminal(BaseTerminal):
+    __slots__ = ()
+
+    def start(self, *, into:tk.Misc=None) -> None:
+        self.run("gnome-terminal -- %command%")
+        self.sep_window:bool = True
+
+
+class ConhostTerminal(BaseTerminal):
+    __slots__ = ()
+
+    def start(self, *, into:tk.Misc=None) -> None:
+        self.run("conhost -- %command%")
+        self.sep_window:bool = True
+
+
 class KonsoleTerminal(BaseTerminal):
     __slots__ = ()
 
-    def start(self, ipc_name:str, *, into:tk.Misc=None):
-        raise NotImplementedError("Not fully implemented. " \
-                                  "Read the comment bellow")
-        args:str = KONSOLE_ARGS.format(ipc_name, SELF_PID)
-        self.run(f"konsole {args}")
-        # Now we have to use `NoTitlebarTk._get_parent` on into
-        #   maybe even on into's tk.Tk and then use
-        #   `NoTitlebarTk._reparent_window(child, parent, x, y)`
-        #   where child is the output from "echo $WINDOWID" when run
-        #   from the slave
+    def start(self, *, into:tk.Misc=None) -> None:
+        raise NotImplementedError("Not implemented yet.")
+        command:str = f"konsole -e %command%"
+        self.sep_window:bool = True
+        self.run(command)
+        # `NoTitlebarTk._reparent_window()` and `echo $WINDOWID`
 
     def ___resize(self, width:int, height:int) -> None:
         raise RuntimeError("https://bugs.kde.org/show_bug.cgi?id=238073")
 
 
+AVAILABLE_TERMS:list[type[BaseTerminal]] = [
+                                             XTermTerminal,
+                                             DebugXTermTerminal,
+                                             KonsoleTerminal, # not working
+                                             GnomeTermTerminal, # sep window
+                                             ConhostTerminal, # sep window
+                                           ]
+
 if IS_UNIX:
     Terminal:type = XTermTerminal
+elif IS_WINDOWS:
+    Terminal:type = ConhostTerminal
 else:
     print('[WARNING]: (Terminal) NotImplementedError')
     Terminal:type = type(None)
@@ -232,5 +279,5 @@ assert issubclass(Terminal, BaseTerminal|None), "You must subclass BaseTerminal.
 if __name__ == "__main__":
     XTERM_DEBUG:bool = True
     term:Terminal = Terminal()
-    term.send_event("run", data=("python3",))
+    term.send("run", data=("python",))
     term.bind("", lambda e: print(e))
